+++ test/integration/adapters/indexeddb.js
@@ -1,5 +1,5 @@
 import testAdapter from '../../unit/adapter'
-import IndexedDB from '../../../lib/adapter/adapters/indexeddb'
+import indexeddbAdapter from '../../../lib/adapter/adapters/indexeddb'
 
 
-testAdapter(IndexedDB, {})
+testAdapter(indexeddbAdapter, {})