+++ test/integration/adapters/webstorage.js
@@ -1,0 +1,5 @@
+import testAdapter from '../../unit/adapter'
+import webStorageAdapter from '../../../lib/adapter/adapters/webstorage'
+
+
+testAdapter(webStorageAdapter, {})